add listing fixes
